Renamed components/application_form to components/generic since the components in that folder is now used by other views. Also changed import paths accordingly in a couple of files.
