+++ src/views/applicant/ApplicationFormView.vue
@@ -4,8 +4,8 @@ import { storeToRefs } from "pinia";
 import { useI18n } from "vue-i18n";
 import { useApplicationStore } from "@/stores/applicationForm";
 import { ApplicationTestId } from "@/util/enums";
-import ItemList from "@/components/application_form/ItemList.vue";
-import PersonalInformation from "@/components/application_form/PersonalInformation.vue";
+import ItemList from "@/components/generic/ItemList.vue";
+import PersonalInformation from "@/components/generic/PersonalInformation.vue";
 const applicationStore = useApplicationStore();
 const { competenceList, availabilityList } = storeToRefs(applicationStore);
 const i18n = useI18n();
